matrix id can be packed into 32bit integer to represent any matrix within lattice of 127x127x127 size, this have to be considered...

+++ olex/xpprocs.cpp
@@ -3410,7 +3410,7 @@ void TMainForm::macSplit(TStrObjList &Cmds, const TParamList &Options, TMacroErr
     TCAtom* CA = Atoms[i];
     if( CA->GetEllipsoid() == NULL )  continue;
     vec3d direction;
-    float Length = 0;
+    double Length = 0;
     lbl = CA->Label();
     if( CA->GetEllipsoid()->GetSX() > CA->GetEllipsoid()->GetSY() )  {
       if( CA->GetEllipsoid()->GetSX() > CA->GetEllipsoid()->GetSZ() )  {
